Moved exceptions to apropriate packages
