+++ backend/src/main/java/pl/edu/pw/ee/backend/api/auth/LogoutServiceImpl.java
@@ -12,8 +12,8 @@ import org.springframework.stereotype.Service;
 import pl.edu.pw.ee.backend.config.jwt.interfaces.JwtService;
 import pl.edu.pw.ee.backend.entities.user.User;
 import pl.edu.pw.ee.backend.entities.user.UserRepository;
-import pl.edu.pw.ee.backend.utils.exceptions.InvalidTokenException;
-import pl.edu.pw.ee.backend.utils.exceptions.UserDoesNotExistException;
+import pl.edu.pw.ee.backend.utils.exceptions.auth.InvalidTokenException;
+import pl.edu.pw.ee.backend.utils.exceptions.auth.UserDoesNotExistException;
 
 import java.util.Optional;
